move kinesis state to new file
