+++ rust/connector/src/kinesis/source/mod.rs
@@ -1,2 +1,3 @@
 mod message;
 mod reader;
+mod state;
